Migrate video tests to videoalpha tests, remove video tests.
